Index type changed

+++ src/BmiService.java
@@ -1,7 +1,7 @@
 public class BmiService {
     double calculate(double heightInMeters, int weightInKilo) {
-        double index;
-        index = (weightInKilo/ (heightInMeters * heightInMeters));
+        int index;
+        index = (int) (weightInKilo / (heightInMeters * heightInMeters));
         return index;
     }
 }
